move pysm operator functions to todmap
